Examples/MAX78000/CNN/UNet-demo/Utility/requirements.txt
ft4222>=1.2.1
Pillow>=8.1.0

